+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 2.2.0 (July 10, 2025)
+* Field `Error Codes for retry` was renamed to `Error Codes for Error Handling Policy` to avoid confusing
+* Added new field `Error Codes to emit as messages`
+* Fixed incorrect placement of application/x-www-form-urlencoded parameters in POST requests
+* Updated `component-commons-library` to version 3.2.2
+* Updated `Sailor` to version to 2.7.5
+
 ## 2.1.0 (November 14, 2024)
 * Deprecated the action and trigger **HTTP Request (Request Library)** - the primary reason for this change is the reliance on the deprecated **request** library. 
 * Introduced a new action and trigger **HTTP Request (Axios Library)**.